find using go routines
speeds up finding by magnitudes. went from an example of 17s to 5s

currently this is tuned to 512 concurrent finds which is still plenty fast
but also causes a bit of cpu load, we may choose to go lower at some point.

up to 2048 result values are cached as to not bottleneck the finding. this
value only increases memory consumption